Extract marketplace data fetching to a store
